# docs/swagger.yaml
openapi: 3.0.1
info:
  title: API tmdb
  contact: {}
  version: 1.0.0
servers:
  - url: http://localhost:3000/
...
paths:
  /api/users:
    get:
      tags:
        - user
      summary: Obtiene la lista de usuarios
      operationId: getUsers
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "c86cef30-383d-49de-bd95-778e309c52b8"
                    name:
                      type: string
                      example: "John"
                    lastname:
                      type: string
                      example: "Doe"
                    email:
                      type: string
                      example: "johndoe@mail.com"
                    createdAt:
                      type: string
                      format: date-time
                      example: "2025-01-30T20:34:34.123Z"
        "400":
          description: No se encontraron usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
              example: []

    post:
      tags:
        - user
      summary: Registra un nuevo usuario
      description: Esta ruta registra un nuevo usuario con los campos `name`, `lastname`, `email` y `password`.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - lastname
                - email
                - password
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del usuario.
                  example: "Michael"
                lastname:
                  type: string
                  description: El apellido del usuario.
                  example: "Jordan"
                email:
                  type: string
                  description: El correo electrónico del usuario. Debe ser válido.
                  example: "michaljordan@mail.com"
                password:
                  type: string
                  description: La contraseña del usuario. Debe cumplir con los requisitos de seguridad (mínimo 8 caracteres, 1 mayúscula, 1 número y 1 carácter especial).
                  example: "Password123!"
      responses:
        "201":
          description: Usuario registrado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario registrado satisfactoriamente."
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "7ead6e2f-bb02-4b6c-8871-084ff2e7b782"
                      name:
                        type: string
                        example: "Michael"
                      lastname:
                        type: string
                        example: "Jordan"
                      email:
                        type: string
                        example: "michaljordan@mail.com"
                      password:
                        type: string
                        example: "$2b$10$dsCFTaY.KTaijAsFklh59uP45KdyarDuitLtqR.RbW86uJixcCqRq"
                      createdAt:
                        type: string
                        example: "2025-01-30T21:02:15.265Z"
                      updatedAt:
                        type: string
                        example: "2025-01-30T21:02:15.265Z"
                      moviesId:
                        type: object
                        example: null
        "400":
          description: Errores de validación. El campo no cumple con los requisitos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errores:
                    type: integer
                    example: 8
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          example: "field"
                        msg:
                          type: string
                          example: "El nombre es requerido."
                        path:
                          type: string
                          example: "name"
                        location:
                          type: string
                          example: "body"

    put:
      tags:
        - user
      summary: Modifica un usuario
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - lastname
                - email
                - password
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del usuario.
                  example: "Michael"
                lastname:
                  type: string
                  description: El apellido del usuario.
                  example: "Jordan"
                email:
                  type: string
                  description: El correo electrónico del usuario. Debe ser válido.
                  example: "michaljordan@mail.com"
                password:
                  type: string
                  description: La contraseña del usuario. Debe cumplir con los requisitos de seguridad (mínimo 8 caracteres, 1 mayúscula, 1 número y 1 carácter especial).
                  example: "Password123!"
      responses:
        "202":
          description: Usuario modificado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario modificado satisfactoriamente."
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "7ead6e2f-bb02-4b6c-8871-084ff2e7b782"
                      name:
                        type: string
                        example: "Michael"
                      lastname:
                        type: string
                        example: "Jordan"
                      email:
                        type: string
                        example: "michaljordan@mail.com"
                      password:
                        type: string
                        example: "$2b$10$dsCFTaY.KTaijAsFklh59uP45KdyarDuitLtqR.RbW86uJixcCqRq"
                      createdAt:
                        type: string
                        example: "2025-01-30T21:02:15.265Z"
                      updatedAt:
                        type: string
                        example: "2025-01-30T21:02:15.265Z"
                      moviesId:
                        type: object
                        example: null
        "404":
          description: Error
  /api/users/user/{id}:
    get:
      tags:
        - user
      summary: Busca un usuario por ID
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "9ad4e065-d6a9-4ad9-a69b-f6972786ebf8"
      responses:
        "200":
          description: Usuario encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "9ad4e065-d6a9-4ad9-a69b-f6972786ebf8"
                  name:
                    type: string
                    example: "John"
                  lastname:
                    type: string
                    example: "Doe"
                  email:
                    type: string
                    example: "johndoe@mail.com"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2025-01-30T20:34:34.123Z"
                  updatedAt:
                    type: string
                    format: date-time
                    example: "2025-01-30T20:35:27.556Z"
                  moviesId:
                    type: array
                    items:
                      type: string
                  favorites:
                    type: array
                    items:
                      type: object
        "400":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "El usuario no existe."

    delete:
      tags:
        - user
      summary: Elimina un usuario
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "7ead6e2f-bb02-4b6c-8871-084ff2e7b782"

      responses:
        "200":
          description: Usuario eliminado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "El usuario ha sido eliminado"
        "400":
          description: Error
  /api/users/login:
    post:
      tags:
        - user
      summary: Registra un nuevo usuario
      description: Esta ruta registra un nuevo usuario con los campos `email` y `password`.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              type: object
              properties:
                email:
                  type: string
                  description: El correo electrónico del usuario. Debe ser válido.
                  example: "michaljordan@mail.com"
                password:
                  type: string
      responses:
        "200":
          description: Login exitoso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: object
                    properties:
                      email:
                        type: string
                        example: "michaljordann@mail.com"
                      name:
                        type: string
                        example: "Michael"
                      lastname:
                        type: string
                        example: "Jordan"
                      id:
                        type: string
                        example: "2a54e5c4-d069-4b88-95fd-7ac52a57fa94"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImVtYWlsIjoibWljaGFsam9yZGFubkBtYWlsLmNvbSIsIm5hbWUiOiJNaWNoYWVsIiwibGFzdG5hbWUiOiJKb3JkYW4iLCJpZCI6IjJhNTRlNWM0LWQwNjktNGI4OC05NWZkLTdhYzUyYTU3ZmE5NCJ9LCJpYXQiOjE3MzgzNjA4MTEsImV4cCI6MTczODM2MTcxMX0.iSvyI3Ey1VtwFEZAKOpoztKd4REw_ZSxFJfZLu2rNZQ"
        "401":
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Credenciales inválidas"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario no encontrado"
...